+++ internal/services/filesharer/internal/stream/stream.go
@@ -7,6 +7,7 @@ import (
 	"errors"
 	"hash"
 	"io"
+	"os"
 
 	"github.com/number571/go-peer/pkg/crypto/hashing"
 	"github.com/number571/go-peer/pkg/encoding"
@@ -31,11 +32,13 @@ var (
 type sStream struct {
 	fContext   context.Context
 	fRetryNum  uint64
+	fTempFile  string
 	fHlfClient hls_filesharer_client.IClient
 	fAliasName string
 	fFileInfo  IFileInfo
 	fBuffer    []byte
 	fPosition  uint64
+	fTempBytes []byte
 	fHasher    hash.Hash
 	fChunkSize uint64
 }
@@ -43,6 +46,7 @@ type sStream struct {
 func BuildStream(
 	pCtx context.Context,
 	pRetryNum uint64,
+	pTempFile string,
 	pHlsClient hls_client.IClient,
 	pAliasName string,
 	pFileInfo IFileInfo,
@@ -51,10 +55,15 @@ func BuildStream(
 	if err != nil {
 		return nil, errors.Join(ErrGetMessageLimit, err)
 	}
-
+	tempBytes, err := os.ReadFile(pTempFile) // nolint: gosec
+	if err != nil {
+		panic(err)
+	}
 	return &sStream{
-		fContext:  pCtx,
-		fRetryNum: pRetryNum,
+		fContext:   pCtx,
+		fRetryNum:  pRetryNum,
+		fTempFile:  pTempFile,
+		fTempBytes: tempBytes,
 		fHlfClient: hls_filesharer_client.NewClient(
 			hls_filesharer_client.NewBuilder(),
 			hls_filesharer_client.NewRequester(pHlsClient),
@@ -67,20 +76,35 @@ func BuildStream(
 }
 
 func (p *sStream) Read(b []byte) (int, error) {
+	select {
+	case <-p.fContext.Done():
+		return 0, io.ErrClosedPipe
+	default:
+	}
+
+	if p.fTempBytes != nil {
+		p.fBuffer = p.fTempBytes
+		p.fTempBytes = nil
+	}
+
 	if len(p.fBuffer) == 0 {
 		chunk, err := p.loadFileChunk()
 		if err != nil {
 			return 0, errors.Join(ErrLoadFileChunk, err)
 		}
-		if _, err := p.fHasher.Write(chunk); err != nil {
+		p.fBuffer = chunk
+		if err := p.appendToTempFile(p.fBuffer); err != nil {
 			return 0, errors.Join(ErrWriteFileChunk, err)
 		}
-		p.fBuffer = chunk
 	}
 
 	n := copy(b, p.fBuffer)
 	p.fBuffer = p.fBuffer[n:]
 	p.fPosition += uint64(n) //nolint:gosec
+
+	if _, err := p.fHasher.Write(b[:n]); err != nil {
+		return 0, errors.Join(ErrWriteFileChunk, err)
+	}
 
 	if p.fPosition < p.fFileInfo.GetSize() {
 		return n, nil
@@ -95,6 +119,12 @@ func (p *sStream) Read(b []byte) (int, error) {
 }
 
 func (p *sStream) Seek(offset int64, whence int) (int64, error) {
+	select {
+	case <-p.fContext.Done():
+		return 0, io.ErrClosedPipe
+	default:
+	}
+
 	var pos int64
 	switch whence {
 	case io.SeekStart:
@@ -130,4 +160,16 @@ func (p *sStream) loadFileChunk() ([]byte, error) {
 		return chunk, nil
 	}
 	return nil, errors.Join(ErrRetryFailed, lastErr)
+}
+
+func (p *sStream) appendToTempFile(chunk []byte) error {
+	f, err := os.OpenFile(p.fTempFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
+	if err != nil {
+		return err
+	}
+	defer func() { _ = f.Close() }()
+	if _, err := f.Write(chunk); err != nil {
+		return err
+	}
+	return nil
 }