+++ internal/services/filesharer/pkg/app/init_storage.go
@@ -2,8 +2,12 @@ package app
 
 import (
 	"os"
+	"path/filepath"
+
+	hls_filesharer_settings "github.com/number571/hidden-lake/internal/services/filesharer/pkg/settings"
 )
 
 func (p *sApp) initStorage() error {
-	return os.MkdirAll(p.fStgPath, 0o777) //nolint:gosec
+	stgPath := filepath.Join(p.fPathTo, hls_filesharer_settings.CPathSTG)
+	return os.MkdirAll(stgPath, 0o777) //nolint:gosec
 }